Gentleman... Models
Assimp integration and a model is drawn! Right now she's power lines
that are pretty damn big.  The draw order should probably be changed to
CW instead of CCW cause most models seem to be CW.

Some further investigation and improvements will need to be made for
texture maps and missing polys